% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.nexus.R
\name{import.nexus}
\alias{import.nexus}
\title{Import the network from a nexus file produced by SplitsTree}
\usage{
import.nexus(nexus)
}
\arguments{
\item{nexus}{the name of a nexus file}
}
\value{
a list of TRANSLATE, VERTICES, EDGES and FROM-TO
}
\description{
\code{import.nexus} reads and import into R the chunks of a nexus file: TRANSLATE, VERTICES and EDGES
}
\examples{
nex = import.nexus(myfile.nexus)
str(nex)

}
\author{
Alessandro Brozzi
}
