% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nexus.R
\name{plot_nexus}
\alias{plot_nexus}
\title{Plot the network}
\usage{
plot_nexus(obj, taxa = NULL, bg, col, cex, pch, text, arrange, col.bg,
  defaults, ylim = NULL, main = NULL, lty = 1, lwd = 1,
  col.segments = "gray")
}
\arguments{
\item{arrange}{can be one of the two: "back" or "front"}

\item{col.bg}{is the color of the box of the text}

\item{defaults}{is a list with names: col, bg, cex, pch, text, arrange}
}
\value{
plot
}
\description{
\code{plot_nexus} plot the network
}
\examples{
nex = import.nexus(myfile.nexus)
plot_nexus(nex)

}
\author{
Alessandro Brozzi
}
